requirements.txt
pdfminer.six==20201018
python-bidi==0.4.2